+++ src/routes/users/+page.svelte
@@ -11,14 +11,15 @@
 <h1>Add New User</h1>
 <form method="POST" action="?/create" use:enhance>
 	<label for="fullName">Name</label>
-	<input type="text" name="fullName" id="fullName" placeholder="J. Random Person" />
+	<input type="text" name="fullName" id="fullName" placeholder="J. Random Person" required />
 	<label for="email">Email</label>
-	<input type="email" name="email" id="email" placeholder="email@example.com" />
+	<input type="email" name="email" id="email" placeholder="email@example.com" required />
 	<Dropdown
 		label="Role"
 		name="role"
 		options={['HACKER', 'ADMIN', 'ORGANIZER', 'MENTOR', 'JUDGE', 'VOLUNTEER', 'SPONSOR']}
 		value=""
+		required
 	/>
 	<button type="submit" value="Create">Create User</button>
 </form>